refactor: move routes to features folder
