+++ server/features/zotero/zotero.test.ts
@@ -3,7 +3,7 @@ import { describe, it } from "node:test";
 
 import request from "supertest";
 
-import { server } from "../server";
+import { server } from "../../server";
 import {
 	type ZoteroItemPatchInput,
 	type ZoteroItemPostInput,
